Inventory service fix #1

+++ ecommerce-microservices/inventory-service/src/main/java/com/toleyko/springboot/inventoryservice/dao/InventoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface InventoryRepository extends JpaRepository<Remainder, Integer> {
+public interface InventoryRepository extends JpaRepository<Remainder, Long> {
     public Optional<Remainder> findByName(String name);
 }
