+++ ecommerce-microservices/inventory-service/src/main/java/com/toleyko/springboot/inventoryservice/service/InventoryService.java
@@ -6,10 +6,10 @@ import com.toleyko.springboot.inventoryservice.handlers.exception.RemainderNotFo
 import java.util.List;
 public interface InventoryService {
     public List<Remainder> getAllRemainders();
-    public Remainder getRemainderById(Integer id) throws RemainderNotFoundException;
+    public Remainder getRemainderById(Long id) throws RemainderNotFoundException;
     public Remainder getRemainderByName(String name) throws RemainderNotFoundException;
-    public Remainder deleteRemainderById(Integer id) throws RemainderNotFoundException;
+    public Remainder deleteRemainderById(Long id) throws RemainderNotFoundException;
     public Remainder saveRemainder(Remainder remainder);
-    public Remainder updateRemainderById(Integer id, Remainder remainder) throws RemainderNotFoundException;
-    public Remainder updateRemainderLeftAmount(Integer id, Integer amount) throws RemainderNotFoundException;
+    public Remainder updateRemainderById(Long id, Remainder remainder) throws RemainderNotFoundException;
+    public Remainder updateRemainderLeftAmount(Long id, Integer amount) throws RemainderNotFoundException;
 }
